+++ .storybook/addons.js
@@ -1,5 +1,5 @@
-import "@storybook/addon-storysource/register";
 import "@storybook/addon-notes/register";
+import "@storybook/addon-storysource/register";
 import '@storybook/addon-actions/register';
 import "@storybook/addon-a11y/register";
 import "@storybook/addon-viewport/register";
