+++ stories/StaticComponents/alertBadge.md
@@ -1,0 +1,16 @@
+
+# alert badge component
+
+&nbsp;
+
+### Example of usage in bourbon
+```
+  {{bourbon-alert-badge message="This feature isn't available on your current plan." }}
+
+```
+&nbsp;
+
+## Properties
+| propertyOptions | type | defaultValue | isRequired | description | options |
+|----------|:----------|:--------------|:------------|:-------------|:------|
+| message | string | null| true| text inside the alert badge||
